modularize further
