+++ .gitignore
@@ -1,3 +1,4 @@
 *.tsv
+*.o
 .vscode/
 Y2-12-7
